@@ -19,72 +19,71 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.3",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.3",
-    "@angular/compiler": "^7.2.3",
-    "@angular/core": "^7.2.3",
-    "@angular/forms": "^7.2.3",
-    "@angular/http": "^7.2.3",
-    "@angular/platform-browser": "^7.2.3",
-    "@angular/platform-browser-dynamic": "^7.2.3",
-    "@angular/router": "^7.2.3",
-    "@fortawesome/fontawesome-free": "^5.1.0",
-    "@ng-bootstrap/ng-bootstrap": "^4.0.4",
-    "@ng-select/ng-select": "^2.19.0",
+    "@angular/animations": "^8.2.8",
+    "@angular/common": "^8.2.8",
+    "@angular/compiler": "^8.2.8",
+    "@angular/core": "^8.2.8",
+    "@angular/forms": "^8.2.8",
+    "@angular/platform-browser": "^8.2.8",
+    "@angular/platform-browser-dynamic": "^8.2.8",
+    "@angular/router": "^8.2.8",
+    "@fortawesome/fontawesome-free": "^5.11.2",
+    "@ng-bootstrap/ng-bootstrap": "^4.2.2",
+    "@ng-select/ng-select": "^2.20.5",
     "@ngx-translate/core": "^11.0.1",
     "@swimlane/ngx-charts": "^12.0.1",
-    "@swimlane/ngx-datatable": "^14.0.0",
-    "angular-highcharts": "^7.2.0",
+    "@swimlane/ngx-datatable": "^16.0.0",
+    "angular-highcharts": "^8.0.0",
     "bootstrap": "^4.1.1",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^2.6.3",
-    "highcharts": "^7.1.1",
+    "core-js": "^2.6.9",
+    "highcharts": "^7.2.0",
     "ng-bootstrap-form-validation": "^4.0.0",
     "ng-push": "^0.2.2",
-    "ngx-gauge": "^1.0.0-beta.10",
+    "ngx-color-picker": "^8.2.0",
+    "ngx-gauge": "^1.0.0-beta.11",
     "ngx-json-viewer": "^2.4.0",
     "ngx-perfect-scrollbar": "^7.2.1",
-    "rxjs": "^6.4.0",
-    "toppy": "^2.3.3",
-    "web-animations-js": "^2.3.1",
+    "rxjs": "^6.5.3",
+    "toppy": "^2.3.4",
+    "web-animations-js": "^2.3.2",
     "zone.js": "^0.9.0"
   },
   "devDependencies": {
-    "@angular/cli": "~7.3.0",
-    "@angular/compiler-cli": "^7.2.3",
-    "@angular/language-service": "^7.2.3",
-    "@angular-devkit/build-angular": "^0.13.0",
+    "@angular-devkit/build-angular": "^0.803.6",
+    "@angular/cli": "~8.3.6",
+    "@angular/compiler-cli": "^8.2.8",
+    "@angular/language-service": "^8.2.8",
     "@biesbjerg/ngx-translate-extract": "^2.3.4",
-    "@ngx-rocket/scripts": "^3.0.0",
-    "@types/jasmine": "^3.3.8",
-    "@types/jasminewd2": "^2.0.3",
-    "@types/node": "^8.10.13",
-    "codelyzer": "^5.0.0",
+    "@ngx-rocket/scripts": "^3.0.4",
+    "@types/jasmine": "^3.4.1",
+    "@types/jasminewd2": "^2.0.7",
+    "@types/node": "^8.10.54",
+    "codelyzer": "^5.1.2",
     "htmlhint": "^0.11.0",
-    "https-proxy-agent": "^2.0.0",
+    "https-proxy-agent": "^2.2.2",
+    "husky": "^1.3.1",
     "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.1.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.2",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
     "karma-junit-reporter": "^1.2.0",
-    "prettier": "^1.16.3",
-    "tslint-config-prettier": "^1.14.0",
-    "stylelint-config-prettier": "^5.1.0",
-    "pretty-quick": "^1.10.0",
-    "husky": "^1.3.1",
+    "prettier": "^1.18.2",
+    "pretty-quick": "^1.11.1",
     "protractor": "~5.4.0",
     "stylelint": "~10.0.1",
+    "stylelint-config-prettier": "^5.3.0",
     "stylelint-config-recommended-scss": "~3.3.0",
     "stylelint-config-standard": "~18.3.0",
     "stylelint-scss": "~3.6.0",
-    "ts-node": "~8.1.0",
+    "ts-node": "^8.4.1",
     "tslint": "~5.16.0",
-    "typescript": "~3.2.0"
+    "tslint-config-prettier": "^1.14.0",
+    "typescript": "~3.5.0"
   },
   "prettier": {
     "singleQuote": true,
