@@ -1,24 +1,22 @@
 import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { CoreModule } from '@app/core';
 import { SharedModule } from '@app/shared';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
 import { NgBootstrapFormValidationModule } from 'ng-bootstrap-form-validation';
+import { PushNotificationsModule } from 'ng-push';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DashboardModule } from './dashboard/dashboard.module';
-import { ShellModule } from './shell/shell.module';
-import { PushNotificationsModule } from 'ng-push';
 import { DeviceByNameComponent } from './dashboard/device-by-name/device-by-name.component';
+import { ShellModule } from './shell/shell.module';
 
 @NgModule({
   imports: [
     BrowserModule,
     //    ServiceWorkerModule.register('./ngsw-worker.js', { enabled: environment.production }),
-    FormsModule,
     HttpClientModule,
     TranslateModule.forRoot(),
     NgBootstrapFormValidationModule.forRoot(),
